Added Create/SetDisabledIconStrip() to TWindowsToolbar (Win32 only)
The default disabled icons of MaxGUIs TWindowsToolbar have a grayish-outline when using alpha channel icons. This commit allows to replace the windows-disabled-icons-approach with a custom iconstrip. There is also a new function added to create a disabled iconstrip out of a given pixmap. The pixmap gets desaturated and alpha-adjusted according to the given parameters.